requirements.txt
requests
beautifulsoup4
lxml
python-dotenv
schedule
tqdm
